ai-agent/requirements.txt
# Install numpy first (compatible with Python 3.13)
numpy>=2.0.0

# Crypto.com AI Agent SDK (will skip numpy constraint)
cryptocom-agent-client

# X402 Paytech Protocol Support
crypto-com-facilitator-client>=1.0.3

# Autonomous Trading Dependencies
vaderSentiment>=3.3.2
schedule>=1.2.0
feedparser>=6.0.10  # NEW: For RSS parsing (CryptoPanic + Google News)

# Additional dependencies
python-dotenv
web3
requests
